Incorporate module.rb methods into root exec_js.rb module definition
